requirements.txt
argparse==1.4.0
beautifulsoup4==4.9.3
dnspython==2.1.0
requests==2.25.1
#socket==0.8.0
urllib3==1.26.7
wafw00f==2.1.0

